+++ frontend/app/(public)/user/checkout/page.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from "react";
-import CheckoutContent from "./CheckoutContent";
+import CheckoutContent from "../../../../components/checkout/CheckoutContent";
 
 export default function CheckoutPage() {
   return (
