+++ frontend/types/checkout.ts
@@ -1,0 +1,7 @@
+import { Course } from "./cart";
+
+export type PaymentMethodType = "WALLET" | "STRIPE" | "PAYPAL";
+
+export interface OrderDetailsProps {
+  courseDetails: Course[];
+}
