+++ frontend/types/wallet.types.ts
@@ -13,3 +13,26 @@ export interface WalletApiResponse<T> {
 	message: string;
 	statusCode: number;
 }
+
+
+
+export interface CreateWalletTopUpRequest {
+  amount: number;
+  paymentMethod: "WALLET" | "STRIPE" | "PAYPAL" | "RAZORPAY";
+}
+
+export interface WalletTopUpResponse {
+  transaction: {
+    id: string;
+    amount: number;
+    status: string;
+    type: string;
+    createdAt: string;
+  };
+  session?: {
+    id: string;
+    url: string;
+    payment_status: string;
+    amount_total: number;
+  };
+}
