+++ frontend/utils/formatPrice.ts
@@ -1,13 +1,3 @@
-export const formatDate = (dateString: string): string => {
-	return new Date(dateString).toLocaleDateString("en-US", {
-		year: "numeric",
-		month: "short",
-		day: "numeric",
-		hour: "2-digit",
-		minute: "2-digit",
-	});
-};
-
 export const formatPrice = (price: number): string => {
 	// Convert to string with 2 decimal places without rounding
 	const parts = price.toString().split(".");
