--- frontend/utils/index.ts
@@ -1,7 +1,0 @@
-export function formatDate(date: string) {
-	return new Date(date).toLocaleDateString("en-US", {
-		year: "numeric",
-		month: "short",
-		day: "numeric",
-	});
-}
